Replace Synology DSM services with buttons (#57352)
