+++ .coveragerc
@@ -1102,9 +1102,12 @@ omit =
     homeassistant/components/synology_chat/notify.py
     homeassistant/components/synology_dsm/__init__.py
     homeassistant/components/synology_dsm/binary_sensor.py
+    homeassistant/components/synology_dsm/button.py
     homeassistant/components/synology_dsm/camera.py
     homeassistant/components/synology_dsm/diagnostics.py
+    homeassistant/components/synology_dsm/common.py
     homeassistant/components/synology_dsm/sensor.py
+    homeassistant/components/synology_dsm/service.py
     homeassistant/components/synology_dsm/switch.py
     homeassistant/components/synology_srm/device_tracker.py
     homeassistant/components/syslog/notify.py
