+++ homeassistant/components/synology_dsm/const.py
@@ -32,7 +32,13 @@ from homeassistant.const import (
 from homeassistant.helpers.entity import EntityCategory, EntityDescription
 
 DOMAIN = "synology_dsm"
-PLATFORMS = [Platform.BINARY_SENSOR, Platform.CAMERA, Platform.SENSOR, Platform.SWITCH]
+PLATFORMS = [
+    Platform.BINARY_SENSOR,
+    Platform.BUTTON,
+    Platform.CAMERA,
+    Platform.SENSOR,
+    Platform.SWITCH,
+]
 COORDINATOR_CAMERAS = "coordinator_cameras"
 COORDINATOR_CENTRAL = "coordinator_central"
 COORDINATOR_SWITCHES = "coordinator_switches"
